+++ src/main/resources/org/moddingx/moonstone/file/icon.png
