--- src/main/scala/io/github/noeppi_noeppi/tools/moonstone/PackConfig.scala
@@ -1,3 +1,0 @@
-package io.github.noeppi_noeppi.tools.moonstone
-
-case class PackConfig(minecraft: String, loader: String)
