+++ src/main/scala/org/moddingx/moonstone/PackConfig.scala
@@ -1,0 +1,3 @@
+package org.moddingx.moonstone
+
+case class PackConfig(minecraft: String, loader: String)
