+++ src/main/scala/org/moddingx/moonstone/util/PendingFuture.scala
@@ -1,0 +1,13 @@
+package org.moddingx.moonstone.util
+
+import java.util.concurrent.{Future, TimeUnit}
+
+object PendingFuture extends Future[Nothing] {
+  override def cancel(interrupt: Boolean): Boolean = false
+  override def isCancelled: Boolean = false
+  override def isDone: Boolean = false
+  override def get(): Nothing = throw new NoSuchElementException
+  override def get(timeout: Long, unit: TimeUnit): Nothing = throw new NoSuchElementException
+  
+  def instance[T]: Future[T] = PendingFuture.asInstanceOf[Future[T]]
+}
